update code for Kanna
