--- DU1/src/main/java/DomainModels/CV.java
@@ -1,64 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package DomainModels;
-
-import java.io.Serializable;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-/**
- *
- * @author yugip
- */
-@Entity
-@Table(name = "chucvu")
-public class CV implements Serializable {
-    @Id
-    private String id;
-    private String ma, ten;
-
-    public CV() {
-    }
-
-    public CV(String ma, String ten) {
-        this.ma = ma;
-        this.ten = ten;
-    }
-    
-    
-
-    public CV(String id, String ma, String ten) {
-        this.id = id;
-        this.ma = ma;
-        this.ten = ten;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getMa() {
-        return ma;
-    }
-
-    public void setMa(String ma) {
-        this.ma = ma;
-    }
-
-    public String getTen() {
-        return ten;
-    }
-
-    public void setTen(String ten) {
-        this.ten = ten;
-    }
-    
-    
-}
