+++ DU1/src/main/java/Services/Impl/NhanVienService.java
@@ -4,8 +4,13 @@
  */
 package Services.Impl;
 
+import DomainModels.ChucVu;
+import DomainModels.NhanVien;
+import Respositories.Impl.NhanVienRespository;
 import Services.IManageNhanVienService;
-import ViewModel.NhanVienViewModel;
+import ViewModels.ChucVuViewModel;
+import ViewModels.NhanVienViewModel;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -14,24 +19,84 @@ import java.util.List;
  */
 public class NhanVienService implements IManageNhanVienService {
 
-    @Override
-    public List<NhanVienViewModel> getListFromDb() {
-        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
+    public final NhanVienRespository nhanVienRespository;
+
+    public NhanVienService() {
+        nhanVienRespository = new NhanVienRespository();
     }
 
     @Override
-    public int them(String maNhanVien) {
-        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
+    public List<NhanVienViewModel> getListNhanVienFromDb(String dieuKien) {
+
+        List<NhanVienViewModel> list = new ArrayList<>();
+
+        if (nhanVienRespository.getAllJoin(dieuKien) != null) {
+
+            for (Object[] nhanVien : nhanVienRespository.getAllJoin(dieuKien)) {
+
+                list.add(new NhanVienViewModel(
+                        ((NhanVien) nhanVien[0]).getIdNhanVien(),
+                        ((NhanVien) nhanVien[0]).getMaNhanVien(),
+                        ((NhanVien) nhanVien[0]).getTaiKhoan(),
+                        ((NhanVien) nhanVien[0]).getTenNhanVien(),
+                        ((NhanVien) nhanVien[0]).getCmnd(),
+                        ((NhanVien) nhanVien[0]).getGioiTinh(),
+                        ((NhanVien) nhanVien[0]).getNgaySinh(),
+                        ((NhanVien) nhanVien[0]).getDiaChi(),
+                        ((NhanVien) nhanVien[0]).getSdt(),
+                        ((NhanVien) nhanVien[0]).getEmail(),
+                        ((NhanVien) nhanVien[0]).getAnh(),
+                        ((NhanVien) nhanVien[0]).getNgayTao(),
+                        new ChucVuViewModel(((ChucVu) nhanVien[1]).getIdChucVu(),
+                                ((ChucVu) nhanVien[1]).getMaChucVu(),
+                                ((ChucVu) nhanVien[1]).getTenChucVu()),
+                        ((NhanVien) nhanVien[0]).getTrangThai()
+                ));
+            }
+            return list;
+        }
+        return null;
     }
 
     @Override
-    public int capNhat(String maNhanVien) {
-        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
+    public Object CD(NhanVienViewModel nhanVienViewModel) {
+        
+        return new NhanVien(
+                nhanVienViewModel.getIdNhanVien(),
+                nhanVienViewModel.getMaNhanVien(),
+                nhanVienViewModel.getTaiKhoan(),
+                nhanVienViewModel.getTenNhanVien(),
+                nhanVienViewModel.getCmnd(),
+                nhanVienViewModel.getGioiTinh(),
+                nhanVienViewModel.getNgaySinh(),
+                nhanVienViewModel.getDiaChi(),
+                nhanVienViewModel.getSdt(),
+                nhanVienViewModel.getEmail(),
+                nhanVienViewModel.getAnh(),
+                nhanVienViewModel.getNgayTao(),
+                new ChucVu(
+                        nhanVienViewModel.getChucVuViewModel().getIdChucVu(),
+                        nhanVienViewModel.getChucVuViewModel().getMaChucVu(),
+                        nhanVienViewModel.getChucVuViewModel().getTenChucVu()
+                ),
+                nhanVienViewModel.getTrangThai()
+        );
+    }
+
+    @Override
+    public int them(NhanVienViewModel nhanVienViewModel) {
+
+        return nhanVienRespository.them((NhanVien) CD(nhanVienViewModel));
+    }
+
+    @Override
+    public int capNhat(NhanVienViewModel nhanVienViewModel) {
+        return nhanVienRespository.capNhat((NhanVien) CD(nhanVienViewModel));
     }
 
     @Override
     public int xoa(String maNhanVien) {
-        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
+        return nhanVienRespository.xoa(maNhanVien);
     }
 
 }
