+++ DU1/src/main/java/ViewModels/NhanVienViewModel.java
@@ -1,0 +1,169 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package ViewModels;
+
+import java.util.Date;
+
+/**
+ *
+ * @author suppe
+ */
+public class NhanVienViewModel {
+
+    private String idNhanVien;
+    private String maNhanVien;
+    private String taiKhoan;
+    private String tenNhanVien;
+    private String cmnd;
+    private String gioiTinh;
+    private Date ngaySinh;
+    private String diaChi;
+    private String sdt;
+    private String email;
+    private String anh;
+    private Date ngayTao;
+    private int trangThai;
+    private ChucVuViewModel chucVuViewModel;
+
+    public NhanVienViewModel() {
+    }
+
+    public NhanVienViewModel(String idNhanVien, String maNhanVien, String taiKhoan, String tenNhanVien, String cmnd, String gioiTinh, Date ngaySinh, String diaChi, String sdt, String email, String anh, Date ngayTao, ChucVuViewModel chucVuViewModel, int trangThai) {
+        this.idNhanVien = idNhanVien;
+        this.maNhanVien = maNhanVien;
+        this.taiKhoan = taiKhoan;
+        this.tenNhanVien = tenNhanVien;
+        this.cmnd = cmnd;
+        this.gioiTinh = gioiTinh;
+        this.ngaySinh = ngaySinh;
+        this.diaChi = diaChi;
+        this.sdt = sdt;
+        this.email = email;
+        this.anh = anh;
+        this.ngayTao = ngayTao;
+        this.chucVuViewModel = chucVuViewModel;
+        this.trangThai = trangThai;
+    }
+
+    public String getIdNhanVien() {
+        return idNhanVien;
+    }
+
+    public void setIdNhanVien(String idNhanVien) {
+        this.idNhanVien = idNhanVien;
+    }
+
+    public String getMaNhanVien() {
+        return maNhanVien;
+    }
+
+    public void setMaNhanVien(String maNhanVien) {
+        this.maNhanVien = maNhanVien;
+    }
+
+    public String getTaiKhoan() {
+        return taiKhoan;
+    }
+
+    public void setTaiKhoan(String taiKhoan) {
+        this.taiKhoan = taiKhoan;
+    }
+
+    public String getTenNhanVien() {
+        return tenNhanVien;
+    }
+
+    public void setTenNhanVien(String tenNhanVien) {
+        this.tenNhanVien = tenNhanVien;
+    }
+
+    public String getCmnd() {
+        return cmnd;
+    }
+
+    public void setCmnd(String cmnd) {
+        this.cmnd = cmnd;
+    }
+
+    public String getGioiTinh() {
+        return gioiTinh;
+    }
+
+    public void setGioiTinh(String gioiTinh) {
+        this.gioiTinh = gioiTinh;
+    }
+
+    public Date getNgaySinh() {
+        return ngaySinh;
+    }
+
+    public void setNgaySinh(Date ngaySinh) {
+        this.ngaySinh = ngaySinh;
+    }
+
+    public String getDiaChi() {
+        return diaChi;
+    }
+
+    public void setDiaChi(String diaChi) {
+        this.diaChi = diaChi;
+    }
+
+    public String getSdt() {
+        return sdt;
+    }
+
+    public void setSdt(String sdt) {
+        this.sdt = sdt;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getAnh() {
+        return anh;
+    }
+
+    public void setAnh(String anh) {
+        this.anh = anh;
+    }
+
+    public Date getNgayTao() {
+        return ngayTao;
+    }
+
+    public void setNgayTao(Date ngayTao) {
+        this.ngayTao = ngayTao;
+    }
+
+    public int getTrangThai() {
+        return trangThai;
+    }
+
+    public void setTrangThai(int trangThai) {
+        this.trangThai = trangThai;
+    }
+
+    public ChucVuViewModel getChucVuViewModel() {
+        return chucVuViewModel;
+    }
+
+    public void setChucVuViewModel(ChucVuViewModel chucVuViewModel) {
+        this.chucVuViewModel = chucVuViewModel;
+    }
+
+    public String trangThai() {
+        if (trangThai == 0) {
+            return "Còn làm";
+        } else {
+            return "Đã nghỉ";
+        }
+    }
+}
