+++ DU1/src/main/java/Views/QuanLyNhanVien.java
@@ -1,0 +1,918 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
+ */
+package Views;
+
+import Respositories.Impl.NhanVienRespository;
+import Services.Impl.ChucVuService;
+import Services.Impl.NhanVienService;
+import Utilities.JframeCheck;
+import ViewModels.ChucVuViewModel;
+import ViewModels.NhanVienViewModel;
+import java.awt.Image;
+import java.awt.Label;
+import java.io.File;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+import javax.swing.DefaultComboBoxModel;
+import javax.swing.ImageIcon;
+import javax.swing.JButton;
+import javax.swing.JFileChooser;
+import javax.swing.JLabel;
+import javax.swing.JOptionPane;
+import javax.swing.table.DefaultTableModel;
+
+/**
+ *
+ * @author suppe
+ */
+public class QuanLyNhanVien extends javax.swing.JFrame {
+
+    private DefaultTableModel dtm;
+
+    private final NhanVienService nhanVienService = new NhanVienService();
+    private final ChucVuService chucVuService = new ChucVuService();
+
+    private final JframeCheck jframeCheck = new JframeCheck();
+
+    private final DefaultComboBoxModel dcb = new DefaultComboBoxModel();
+
+    private String duongDanThuMucAnh;
+    
+    private final List<Object> jtext = new ArrayList<>();
+
+    /**
+     * Creates new form QuanLyNhanVien
+     */
+    public QuanLyNhanVien() {
+        initComponents();
+        getIconMenu(buttonTrangChu, "icon\\Images\\Home.png");
+        getIconMenu(buttonBanHang, "icon\\Images\\Basket.png");
+        getIconMenu(buttonNhanVien, "icon\\Images\\User.png");
+        getIconMenu(buttonSanPham, "icon\\Images\\Label.png");
+        getIconMenu(buttonKhachHang, "icon\\Images\\User group.png");
+        getIconMenu(buttonKhuyenMai, "icon\\Images\\Free.png");
+        getIconMenu(buttonThongKe, "icon\\Images\\Diagram.png");
+        getIconMenu(buttonLichSu, "icon\\Images\\Clock.png");
+        getIconMenu(buttonDoiMatKhau, "icon\\Images\\Refresh.png");
+        getIconMenu(buttonDangXuat, "icon\\Images\\Open door.png");
+        getImageThemAvatar(labelThemAnh, duongDanThuMucAnh);
+
+        loadCombobox();
+        loadTable();
+    }
+
+    public void getIconMenu(JButton bt, String dd) {
+        Image image = new ImageIcon(dd).getImage().getScaledInstance(24, 24, 0);
+        bt.setIcon(new ImageIcon(image));
+    }
+
+    public void getImageThemAvatar(JLabel labelThemAnh, String duongDan) {
+        Image image = new ImageIcon(duongDan).getImage().getScaledInstance(labelThemAnh.getWidth(), labelThemAnh.getHeight(), 0);
+        labelThemAnh.setIcon(new ImageIcon(image));
+    }
+
+    public void loadCombobox() {
+//        if (nhanVienService.getChucVu(null) != null) {
+        for (ChucVuViewModel chucVuViewModel : chucVuService.getAllLoad()) {
+            dcb.addElement(chucVuViewModel);
+        }
+        comboboxChucVu.setModel(dcb);
+    }
+//    }
+
+    public void loadTable() {
+        dtm = (DefaultTableModel) tableHienThi.getModel();
+        dtm.setRowCount(0);
+        if (nhanVienService.getListNhanVienFromDb(null) != null) {
+            List<NhanVienViewModel> listNhanVien = nhanVienService.getListNhanVienFromDb("");
+            for (NhanVienViewModel nhanVienViewModel : listNhanVien) {
+                dtm.addRow(
+                        new Object[]{
+                            nhanVienViewModel.getIdNhanVien(),
+                            nhanVienViewModel.getMaNhanVien(),
+                            nhanVienViewModel.getTaiKhoan(),
+                            nhanVienViewModel.getTenNhanVien(),
+                            nhanVienViewModel.getCmnd(),
+                            nhanVienViewModel.getGioiTinh(),
+                            nhanVienViewModel.getNgaySinh(),
+                            nhanVienViewModel.getDiaChi(),
+                            nhanVienViewModel.getSdt(),
+                            nhanVienViewModel.getEmail(),
+                            nhanVienViewModel.getAnh(),
+                            nhanVienViewModel.getNgayTao(),
+                            nhanVienViewModel.getChucVuViewModel(),
+                            nhanVienViewModel.trangThai()
+                        });
+            }
+        }
+    }
+    
+    public void clear() {
+        txtMaNhanVien.setText("");
+        txtTenNhanVien.setText("");
+        txtTenTaiKhoan.setText("");
+        txtSdt.setText("");
+        txtDiaChi.setText("");
+        txtCmnd.setText("");
+        labelThemAnh.setIcon(null);
+        txtEmail.setText("");
+        dateNgaySinh.setDate(null);
+        dateNgayTao.setDate(null);
+        
+    }
+
+    private NhanVienViewModel getDataNhanVien(String nhanVien) {
+        int index = tableHienThi.getSelectedRow();
+        String gioiTinh = null;
+        int trangThai = 0;
+        if (radioNam.isSelected()) {
+            gioiTinh = "Nam";
+        } else {
+            gioiTinh = "Nữ";
+        }
+        if (comboboxTrangThai.getSelectedItem().equals("Còn làm")) {
+            trangThai = 0;
+        } else {
+            trangThai = 1;
+        }
+        if (nhanVien.equals("capNhat")) {
+            System.out.println(tableHienThi.getValueAt(index, 0).toString() + "abc");
+            return new NhanVienViewModel(
+                    tableHienThi.getValueAt(index, 0).toString(),
+                    txtMaNhanVien.getText(),
+                    txtTenTaiKhoan.getText(),
+                    txtTenNhanVien.getText(),
+                    txtCmnd.getText(),
+                    gioiTinh,
+                    dateNgaySinh.getDate(),
+                    txtDiaChi.getText(),
+                    txtSdt.getText(),
+                    txtEmail.getText(),
+                    duongDanThuMucAnh,
+                    dateNgayTao.getDate(),
+                    (ChucVuViewModel) dcb.getSelectedItem(),
+                    trangThai
+            );
+        }
+        
+        System.out.println(duongDanThuMucAnh);
+        return new NhanVienViewModel(
+                jframeCheck.createID().toString(),
+                txtMaNhanVien.getText().trim(),
+                txtTenTaiKhoan.getText().trim(),
+                txtTenNhanVien.getText().trim(),
+                txtCmnd.getText().trim(),
+                gioiTinh,
+                dateNgaySinh.getDate(),
+                txtDiaChi.getText().trim(),
+                txtSdt.getText().trim(),
+                txtEmail.getText().trim(),
+                duongDanThuMucAnh,
+                dateNgayTao.getDate(),
+                (ChucVuViewModel) dcb.getSelectedItem(),
+                trangThai
+        );
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        buttonGroup1 = new javax.swing.ButtonGroup();
+        jPanel1 = new javax.swing.JPanel();
+        jPanel3 = new javax.swing.JPanel();
+        jPanel4 = new javax.swing.JPanel();
+        jLabel1 = new javax.swing.JLabel();
+        jLabel2 = new javax.swing.JLabel();
+        jLabel3 = new javax.swing.JLabel();
+        jLabel4 = new javax.swing.JLabel();
+        buttonDangXuat = new javax.swing.JButton();
+        buttonDoiMatKhau = new javax.swing.JButton();
+        buttonLichSu = new javax.swing.JButton();
+        buttonThongKe = new javax.swing.JButton();
+        buttonKhuyenMai = new javax.swing.JButton();
+        buttonKhachHang = new javax.swing.JButton();
+        buttonSanPham = new javax.swing.JButton();
+        buttonNhanVien = new javax.swing.JButton();
+        buttonBanHang = new javax.swing.JButton();
+        buttonTrangChu = new javax.swing.JButton();
+        jPanel2 = new javax.swing.JPanel();
+        jLabel5 = new javax.swing.JLabel();
+        jPanel5 = new javax.swing.JPanel();
+        jPanel8 = new javax.swing.JPanel();
+        jTabbedPane2 = new javax.swing.JTabbedPane();
+        jPanel6 = new javax.swing.JPanel();
+        jLabel7 = new javax.swing.JLabel();
+        jLabel8 = new javax.swing.JLabel();
+        txtMaNhanVien = new javax.swing.JTextField();
+        txtTenNhanVien = new javax.swing.JTextField();
+        txtCmnd = new javax.swing.JTextField();
+        jLabel9 = new javax.swing.JLabel();
+        radioNam = new javax.swing.JRadioButton();
+        jLabel10 = new javax.swing.JLabel();
+        radioNu = new javax.swing.JRadioButton();
+        dateNgaySinh = new com.toedter.calendar.JDateChooser();
+        jLabel11 = new javax.swing.JLabel();
+        txtSdt = new javax.swing.JTextField();
+        jLabel12 = new javax.swing.JLabel();
+        jLabel13 = new javax.swing.JLabel();
+        txtTenTaiKhoan = new javax.swing.JTextField();
+        dateNgayTao = new com.toedter.calendar.JDateChooser();
+        jLabel14 = new javax.swing.JLabel();
+        txtEmail = new javax.swing.JTextField();
+        jLabel15 = new javax.swing.JLabel();
+        comboboxChucVu = new javax.swing.JComboBox<>();
+        jLabel16 = new javax.swing.JLabel();
+        jLabel17 = new javax.swing.JLabel();
+        comboboxTrangThai = new javax.swing.JComboBox<>();
+        jLabel18 = new javax.swing.JLabel();
+        txtDiaChi = new javax.swing.JTextField();
+        panelThemAnh = new javax.swing.JPanel();
+        labelThemAnh = new javax.swing.JLabel();
+        buttonThem = new javax.swing.JButton();
+        buttonCapNhat = new javax.swing.JButton();
+        buttonXoa = new javax.swing.JButton();
+        jScrollPane1 = new javax.swing.JScrollPane();
+        tableHienThi = new javax.swing.JTable();
+        jButton1 = new javax.swing.JButton();
+        jLabel6 = new javax.swing.JLabel();
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+        setBackground(new java.awt.Color(255, 255, 255));
+        setResizable(false);
+
+        jPanel1.setBackground(new java.awt.Color(238, 235, 155));
+        jPanel1.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
+
+        jPanel3.setBackground(new java.awt.Color(238, 235, 155));
+
+        jPanel4.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
+
+        javax.swing.GroupLayout jPanel4Layout = new javax.swing.GroupLayout(jPanel4);
+        jPanel4.setLayout(jPanel4Layout);
+        jPanel4Layout.setHorizontalGroup(
+            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 85, Short.MAX_VALUE)
+        );
+        jPanel4Layout.setVerticalGroup(
+            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 100, Short.MAX_VALUE)
+        );
+
+        jLabel1.setText("Tên nhân viên:");
+
+        jLabel2.setText("Chức vụ:");
+
+        javax.swing.GroupLayout jPanel3Layout = new javax.swing.GroupLayout(jPanel3);
+        jPanel3.setLayout(jPanel3Layout);
+        jPanel3Layout.setHorizontalGroup(
+            jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel3Layout.createSequentialGroup()
+                .addGap(10, 10, 10)
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGroup(jPanel3Layout.createSequentialGroup()
+                        .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jLabel1)
+                            .addComponent(jLabel2))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jLabel3, javax.swing.GroupLayout.DEFAULT_SIZE, 97, Short.MAX_VALUE)
+                            .addComponent(jLabel4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))))
+                .addContainerGap())
+        );
+        jPanel3Layout.setVerticalGroup(
+            jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel3Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jPanel4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(18, 18, 18)
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel1)
+                    .addComponent(jLabel3))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel2)
+                    .addComponent(jLabel4))
+                .addGap(10, 10, 10))
+        );
+
+        buttonDangXuat.setBackground(new java.awt.Color(238, 235, 155));
+        buttonDangXuat.setText("Đăng xuất");
+        buttonDangXuat.setBorder(null);
+        buttonDangXuat.setBorderPainted(false);
+        buttonDangXuat.setContentAreaFilled(false);
+
+        buttonDoiMatKhau.setBackground(new java.awt.Color(238, 235, 155));
+        buttonDoiMatKhau.setText("Đổi mật khẩu");
+        buttonDoiMatKhau.setBorder(null);
+        buttonDoiMatKhau.setBorderPainted(false);
+        buttonDoiMatKhau.setContentAreaFilled(false);
+
+        buttonLichSu.setBackground(new java.awt.Color(238, 235, 155));
+        buttonLichSu.setText("Lịch sử");
+        buttonLichSu.setBorder(null);
+        buttonLichSu.setBorderPainted(false);
+        buttonLichSu.setContentAreaFilled(false);
+
+        buttonThongKe.setBackground(new java.awt.Color(238, 235, 155));
+        buttonThongKe.setText("Thống kê");
+        buttonThongKe.setBorder(null);
+        buttonThongKe.setBorderPainted(false);
+        buttonThongKe.setContentAreaFilled(false);
+
+        buttonKhuyenMai.setBackground(new java.awt.Color(238, 235, 155));
+        buttonKhuyenMai.setText("Khuyến mãi");
+        buttonKhuyenMai.setBorder(null);
+        buttonKhuyenMai.setBorderPainted(false);
+        buttonKhuyenMai.setContentAreaFilled(false);
+
+        buttonKhachHang.setBackground(new java.awt.Color(238, 235, 155));
+        buttonKhachHang.setText("Khách hàng");
+        buttonKhachHang.setBorder(null);
+        buttonKhachHang.setBorderPainted(false);
+        buttonKhachHang.setContentAreaFilled(false);
+
+        buttonSanPham.setBackground(new java.awt.Color(238, 235, 155));
+        buttonSanPham.setText("Sản phẩm");
+        buttonSanPham.setBorder(null);
+        buttonSanPham.setBorderPainted(false);
+        buttonSanPham.setContentAreaFilled(false);
+
+        buttonNhanVien.setBackground(new java.awt.Color(238, 235, 155));
+        buttonNhanVien.setText("Nhân Viên");
+        buttonNhanVien.setBorder(null);
+        buttonNhanVien.setBorderPainted(false);
+        buttonNhanVien.setContentAreaFilled(false);
+
+        buttonBanHang.setBackground(new java.awt.Color(238, 235, 155));
+        buttonBanHang.setText("Bán hàng");
+        buttonBanHang.setBorder(null);
+        buttonBanHang.setBorderPainted(false);
+        buttonBanHang.setContentAreaFilled(false);
+
+        buttonTrangChu.setBackground(new java.awt.Color(238, 235, 155));
+        buttonTrangChu.setText("Trang chủ");
+        buttonTrangChu.setBorder(null);
+        buttonTrangChu.setBorderPainted(false);
+        buttonTrangChu.setContentAreaFilled(false);
+
+        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
+        jPanel1.setLayout(jPanel1Layout);
+        jPanel1Layout.setHorizontalGroup(
+            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(jPanel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(buttonBanHang, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(buttonNhanVien, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(buttonSanPham, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(buttonKhachHang, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(buttonKhuyenMai, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(buttonThongKe, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(buttonLichSu, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(buttonDoiMatKhau, javax.swing.GroupLayout.DEFAULT_SIZE, 196, Short.MAX_VALUE)
+            .addComponent(buttonDangXuat, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(buttonTrangChu, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        jPanel1Layout.setVerticalGroup(
+            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel1Layout.createSequentialGroup()
+                .addComponent(jPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonTrangChu, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonBanHang, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonNhanVien, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonSanPham, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonKhachHang, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonKhuyenMai, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonThongKe, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonLichSu, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonDoiMatKhau, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(buttonDangXuat, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+
+        jPanel2.setBackground(new java.awt.Color(255, 255, 255));
+        jPanel2.setBorder(javax.swing.BorderFactory.createEtchedBorder());
+
+        jLabel5.setText("Chào mừng tất cả mọi người đến với cửa hàng bán phụ kiện trang sức nữ! Chúc mọi người một ngày mới tốt lành");
+
+        javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
+        jPanel2.setLayout(jPanel2Layout);
+        jPanel2Layout.setHorizontalGroup(
+            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel2Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jLabel5)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+        jPanel2Layout.setVerticalGroup(
+            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(jLabel5)
+                .addContainerGap())
+        );
+
+        jPanel5.setBackground(new java.awt.Color(255, 255, 255));
+
+        jPanel8.setBackground(new java.awt.Color(255, 255, 255));
+
+        jTabbedPane2.setBackground(new java.awt.Color(255, 255, 255));
+        jTabbedPane2.setBorder(javax.swing.BorderFactory.createEtchedBorder());
+
+        jPanel6.setBackground(new java.awt.Color(255, 255, 255));
+
+        jLabel7.setText("Mã nhân viên:");
+
+        jLabel8.setText("Tên nhân viên:");
+
+        jLabel9.setText("CMND:");
+
+        buttonGroup1.add(radioNam);
+        radioNam.setText("Nam");
+
+        jLabel10.setText("Giới tính:");
+
+        buttonGroup1.add(radioNu);
+        radioNu.setText("Nữ");
+
+        dateNgaySinh.setBackground(new java.awt.Color(255, 255, 255));
+        dateNgaySinh.setDateFormatString("dd/MM/yyyy");
+
+        jLabel11.setText("Ngày sinh:");
+
+        jLabel12.setText("Số điện thoại:");
+
+        jLabel13.setText("Tên tài khoản:");
+
+        dateNgayTao.setBackground(new java.awt.Color(255, 255, 255));
+        dateNgayTao.setDateFormatString("dd/MM/yyyy");
+
+        jLabel14.setText("Ngày tạo:");
+
+        jLabel15.setText("Email:");
+
+        jLabel16.setText("Chức vụ:");
+
+        jLabel17.setText("Trạng thái:");
+
+        comboboxTrangThai.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Còn làm", "Đã nghỉ" }));
+
+        jLabel18.setText("Địa chỉ:");
+
+        labelThemAnh.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                labelThemAnhMouseClicked(evt);
+            }
+        });
+
+        javax.swing.GroupLayout panelThemAnhLayout = new javax.swing.GroupLayout(panelThemAnh);
+        panelThemAnh.setLayout(panelThemAnhLayout);
+        panelThemAnhLayout.setHorizontalGroup(
+            panelThemAnhLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(labelThemAnh, javax.swing.GroupLayout.DEFAULT_SIZE, 120, Short.MAX_VALUE)
+        );
+        panelThemAnhLayout.setVerticalGroup(
+            panelThemAnhLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(labelThemAnh, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 160, Short.MAX_VALUE)
+        );
+
+        buttonThem.setText("Lưu");
+        buttonThem.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                buttonThemActionPerformed(evt);
+            }
+        });
+
+        buttonCapNhat.setText("Cập nhật");
+        buttonCapNhat.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                buttonCapNhatActionPerformed(evt);
+            }
+        });
+
+        buttonXoa.setText("Xóa");
+        buttonXoa.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                buttonXoaActionPerformed(evt);
+            }
+        });
+
+        tableHienThi.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+                {null, null, null, null, null, null, null, null, null, null, null, null, null, null},
+                {null, null, null, null, null, null, null, null, null, null, null, null, null, null},
+                {null, null, null, null, null, null, null, null, null, null, null, null, null, null},
+                {null, null, null, null, null, null, null, null, null, null, null, null, null, null}
+            },
+            new String [] {
+                "ID", "Mã", "Tài khoản", "Tên", "CMND", "Giới tính", "Ngày sinh", "Địa chỉ", "Số điện thoại", "Email", "Ảnh", "Ngày vào làm", "Chức vụ", "Trạng thái"
+            }
+        ));
+        tableHienThi.setGridColor(new java.awt.Color(255, 255, 255));
+        tableHienThi.setSelectionBackground(new java.awt.Color(51, 102, 255));
+        tableHienThi.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                tableHienThiMouseClicked(evt);
+            }
+        });
+        jScrollPane1.setViewportView(tableHienThi);
+        if (tableHienThi.getColumnModel().getColumnCount() > 0) {
+            tableHienThi.getColumnModel().getColumn(0).setMinWidth(0);
+            tableHienThi.getColumnModel().getColumn(0).setMaxWidth(0);
+            tableHienThi.getColumnModel().getColumn(1).setMinWidth(50);
+            tableHienThi.getColumnModel().getColumn(1).setMaxWidth(50);
+            tableHienThi.getColumnModel().getColumn(4).setMinWidth(100);
+            tableHienThi.getColumnModel().getColumn(4).setMaxWidth(100);
+            tableHienThi.getColumnModel().getColumn(5).setMinWidth(60);
+            tableHienThi.getColumnModel().getColumn(5).setMaxWidth(60);
+            tableHienThi.getColumnModel().getColumn(10).setMinWidth(0);
+            tableHienThi.getColumnModel().getColumn(10).setMaxWidth(0);
+        }
+
+        jButton1.setText("Clear");
+        jButton1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton1ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel6Layout = new javax.swing.GroupLayout(jPanel6);
+        jPanel6.setLayout(jPanel6Layout);
+        jPanel6Layout.setHorizontalGroup(
+            jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel6Layout.createSequentialGroup()
+                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jPanel6Layout.createSequentialGroup()
+                        .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(jPanel6Layout.createSequentialGroup()
+                                .addGap(80, 80, 80)
+                                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                    .addComponent(jLabel18)
+                                    .addComponent(jLabel8)
+                                    .addComponent(jLabel9)
+                                    .addComponent(jLabel10)
+                                    .addComponent(jLabel11)
+                                    .addComponent(jLabel12))
+                                .addGap(18, 18, 18)
+                                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                    .addGroup(jPanel6Layout.createSequentialGroup()
+                                        .addComponent(radioNam)
+                                        .addGap(18, 18, 18)
+                                        .addComponent(radioNu))
+                                    .addComponent(txtCmnd, javax.swing.GroupLayout.DEFAULT_SIZE, 200, Short.MAX_VALUE)
+                                    .addComponent(txtTenNhanVien, javax.swing.GroupLayout.DEFAULT_SIZE, 200, Short.MAX_VALUE)
+                                    .addComponent(dateNgaySinh, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addComponent(txtSdt)
+                                    .addComponent(txtDiaChi))
+                                .addGap(60, 60, 60)
+                                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addGroup(jPanel6Layout.createSequentialGroup()
+                                        .addComponent(jLabel7)
+                                        .addGap(18, 18, 18)
+                                        .addComponent(txtMaNhanVien, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                    .addGroup(jPanel6Layout.createSequentialGroup()
+                                        .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                            .addComponent(jLabel17)
+                                            .addComponent(jLabel13)
+                                            .addComponent(jLabel14)
+                                            .addComponent(jLabel15)
+                                            .addComponent(jLabel16))
+                                        .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                            .addGroup(jPanel6Layout.createSequentialGroup()
+                                                .addGap(19, 19, 19)
+                                                .addComponent(txtTenTaiKhoan, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                            .addGroup(jPanel6Layout.createSequentialGroup()
+                                                .addGap(18, 18, 18)
+                                                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                                    .addComponent(dateNgayTao, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                    .addComponent(txtEmail, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                            .addGroup(jPanel6Layout.createSequentialGroup()
+                                                .addGap(18, 18, 18)
+                                                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                                    .addComponent(comboboxTrangThai, 0, 130, Short.MAX_VALUE)
+                                                    .addComponent(comboboxChucVu, 0, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))))
+                                        .addGap(60, 60, 60)
+                                        .addComponent(panelThemAnh, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                            .addGroup(jPanel6Layout.createSequentialGroup()
+                                .addGap(273, 273, 273)
+                                .addComponent(buttonThem, javax.swing.GroupLayout.PREFERRED_SIZE, 80, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(70, 70, 70)
+                                .addComponent(buttonCapNhat, javax.swing.GroupLayout.PREFERRED_SIZE, 80, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(70, 70, 70)
+                                .addComponent(buttonXoa, javax.swing.GroupLayout.PREFERRED_SIZE, 80, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(46, 46, 46)
+                                .addComponent(jButton1)))
+                        .addGap(0, 52, Short.MAX_VALUE))
+                    .addGroup(jPanel6Layout.createSequentialGroup()
+                        .addContainerGap()
+                        .addComponent(jScrollPane1)))
+                .addContainerGap())
+        );
+        jPanel6Layout.setVerticalGroup(
+            jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel6Layout.createSequentialGroup()
+                .addGap(40, 40, 40)
+                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel7)
+                    .addComponent(txtMaNhanVien, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(txtTenNhanVien, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel8))
+                .addGap(30, 30, 30)
+                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jPanel6Layout.createSequentialGroup()
+                        .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addGroup(jPanel6Layout.createSequentialGroup()
+                                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                    .addComponent(jLabel13)
+                                    .addComponent(txtTenTaiKhoan, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addComponent(txtCmnd, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addComponent(jLabel9))
+                                .addGap(30, 30, 30)
+                                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                    .addComponent(txtEmail, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addComponent(jLabel15)
+                                    .addComponent(radioNam)
+                                    .addComponent(radioNu)
+                                    .addComponent(jLabel10))
+                                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addGroup(jPanel6Layout.createSequentialGroup()
+                                        .addGap(30, 30, 30)
+                                        .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                            .addComponent(jLabel14)
+                                            .addComponent(dateNgayTao, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel6Layout.createSequentialGroup()
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                        .addComponent(jLabel11))))
+                            .addComponent(dateNgaySinh, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addGap(30, 30, 30)
+                        .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(comboboxChucVu, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel16)
+                            .addComponent(txtSdt, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel12)))
+                    .addComponent(panelThemAnh, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGap(30, 30, 30)
+                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel17)
+                    .addComponent(comboboxTrangThai, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(txtDiaChi, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel18))
+                .addGap(40, 40, 40)
+                .addGroup(jPanel6Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(buttonThem, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(buttonCapNhat, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(buttonXoa, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jButton1))
+                .addGap(40, 40, 40)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 217, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+
+        jTabbedPane2.addTab("Thông tin nhân viên", jPanel6);
+
+        jLabel6.setFont(new java.awt.Font("Segoe UI", 1, 18)); // NOI18N
+        jLabel6.setText("Quản lý nhân viên");
+
+        javax.swing.GroupLayout jPanel8Layout = new javax.swing.GroupLayout(jPanel8);
+        jPanel8.setLayout(jPanel8Layout);
+        jPanel8Layout.setHorizontalGroup(
+            jPanel8Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel8Layout.createSequentialGroup()
+                .addGap(390, 390, 390)
+                .addComponent(jLabel6)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+            .addGroup(jPanel8Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jTabbedPane2)
+                .addContainerGap())
+        );
+        jPanel8Layout.setVerticalGroup(
+            jPanel8Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel8Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jLabel6)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jTabbedPane2))
+        );
+
+        javax.swing.GroupLayout jPanel5Layout = new javax.swing.GroupLayout(jPanel5);
+        jPanel5.setLayout(jPanel5Layout);
+        jPanel5Layout.setHorizontalGroup(
+            jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(jPanel8, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        jPanel5Layout.setVerticalGroup(
+            jPanel5Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel5Layout.createSequentialGroup()
+                .addGap(0, 0, 0)
+                .addComponent(jPanel8, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 0, 0))
+        );
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 0, 0)
+                .addComponent(jPanel5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+            .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jPanel5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void tableHienThiMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_tableHienThiMouseClicked
+        int index = tableHienThi.getSelectedRow();
+        Image image = new ImageIcon(tableHienThi.getValueAt(index, 10).toString()).getImage().getScaledInstance(labelThemAnh.getWidth(), labelThemAnh.getHeight(), 0);        
+        
+        txtMaNhanVien.setText((String) tableHienThi.getValueAt(index, 1));
+        txtTenTaiKhoan.setText((String) tableHienThi.getValueAt(index, 2));
+        txtTenNhanVien.setText((String) tableHienThi.getValueAt(index, 3));
+        txtCmnd.setText((String) tableHienThi.getValueAt(index, 4));
+        if (tableHienThi.getValueAt(index, 5).equals("Nam")) {
+            radioNam.setSelected(true);
+        } else {
+            radioNu.setSelected(true);
+        }
+        dateNgaySinh.setDate((Date) tableHienThi.getValueAt(index, 6));
+        txtDiaChi.setText((String) tableHienThi.getValueAt(index, 7));
+        txtSdt.setText((String) tableHienThi.getValueAt(index, 8));
+        txtEmail.setText((String) tableHienThi.getValueAt(index, 9));
+        labelThemAnh.setIcon(new ImageIcon(image));
+        dateNgayTao.setDate((Date) tableHienThi.getValueAt(index, 11));
+        comboboxChucVu.setSelectedItem(tableHienThi.getValueAt(index, 12));
+        comboboxTrangThai.setSelectedItem((String) tableHienThi.getValueAt(index, 13));
+    }//GEN-LAST:event_tableHienThiMouseClicked
+
+    private void buttonThemActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_buttonThemActionPerformed
+        if (nhanVienService.them(getDataNhanVien("")) == 1) {
+            return;
+        }
+        clear();
+        loadTable();
+    }//GEN-LAST:event_buttonThemActionPerformed
+
+    private void labelThemAnhMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_labelThemAnhMouseClicked
+        JFileChooser chonAnh = new JFileChooser("icon\\Icon");
+
+        chonAnh.showOpenDialog(this);
+        File fileAnh = chonAnh.getSelectedFile();
+        Image anh = new ImageIcon("icon\\Icon\\" + fileAnh.getName()).getImage().getScaledInstance(labelThemAnh.getWidth(), panelThemAnh.getHeight(), 0);
+        labelThemAnh.setIcon(new ImageIcon(anh));
+        duongDanThuMucAnh = "icon\\Icon\\" + fileAnh.getName();
+    }//GEN-LAST:event_labelThemAnhMouseClicked
+
+    private void buttonCapNhatActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_buttonCapNhatActionPerformed
+        if (nhanVienService.capNhat(getDataNhanVien("capNhat")) == 1) {
+            return;
+        }
+        loadTable();
+    }//GEN-LAST:event_buttonCapNhatActionPerformed
+
+    private void buttonXoaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_buttonXoaActionPerformed
+        int index = tableHienThi.getSelectedRow();
+        if (index < 0) {
+            JOptionPane.showMessageDialog(this, "Chưa chọn dòng cần xóa tại bảng!");
+        } else {
+            int x = JOptionPane.showConfirmDialog(this, "Bạn có chắc muốn xóa không?", "Thông báo:", JOptionPane.YES_NO_OPTION);
+            if (x == JOptionPane.YES_OPTION) {
+                nhanVienService.xoa(tableHienThi.getValueAt(index, 0).toString());
+                JOptionPane.showMessageDialog(this, "Xóa thành công!");
+            } else {
+                JOptionPane.showMessageDialog(this, "Xóa thất bại");
+            }
+            clear();
+            loadTable();
+        }
+    }//GEN-LAST:event_buttonXoaActionPerformed
+
+    private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
+        clear();
+        
+    }//GEN-LAST:event_jButton1ActionPerformed
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(QuanLyNhanVien.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(QuanLyNhanVien.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(QuanLyNhanVien.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(QuanLyNhanVien.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            public void run() {
+                new QuanLyNhanVien().setVisible(true);
+            }
+        });
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton buttonBanHang;
+    private javax.swing.JButton buttonCapNhat;
+    private javax.swing.JButton buttonDangXuat;
+    private javax.swing.JButton buttonDoiMatKhau;
+    private javax.swing.ButtonGroup buttonGroup1;
+    private javax.swing.JButton buttonKhachHang;
+    private javax.swing.JButton buttonKhuyenMai;
+    private javax.swing.JButton buttonLichSu;
+    private javax.swing.JButton buttonNhanVien;
+    private javax.swing.JButton buttonSanPham;
+    private javax.swing.JButton buttonThem;
+    private javax.swing.JButton buttonThongKe;
+    private javax.swing.JButton buttonTrangChu;
+    private javax.swing.JButton buttonXoa;
+    private javax.swing.JComboBox<String> comboboxChucVu;
+    private javax.swing.JComboBox<String> comboboxTrangThai;
+    private com.toedter.calendar.JDateChooser dateNgaySinh;
+    private com.toedter.calendar.JDateChooser dateNgayTao;
+    private javax.swing.JButton jButton1;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JLabel jLabel10;
+    private javax.swing.JLabel jLabel11;
+    private javax.swing.JLabel jLabel12;
+    private javax.swing.JLabel jLabel13;
+    private javax.swing.JLabel jLabel14;
+    private javax.swing.JLabel jLabel15;
+    private javax.swing.JLabel jLabel16;
+    private javax.swing.JLabel jLabel17;
+    private javax.swing.JLabel jLabel18;
+    private javax.swing.JLabel jLabel2;
+    private javax.swing.JLabel jLabel3;
+    private javax.swing.JLabel jLabel4;
+    private javax.swing.JLabel jLabel5;
+    private javax.swing.JLabel jLabel6;
+    private javax.swing.JLabel jLabel7;
+    private javax.swing.JLabel jLabel8;
+    private javax.swing.JLabel jLabel9;
+    private javax.swing.JPanel jPanel1;
+    private javax.swing.JPanel jPanel2;
+    private javax.swing.JPanel jPanel3;
+    private javax.swing.JPanel jPanel4;
+    private javax.swing.JPanel jPanel5;
+    private javax.swing.JPanel jPanel6;
+    private javax.swing.JPanel jPanel8;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JTabbedPane jTabbedPane2;
+    private javax.swing.JLabel labelThemAnh;
+    private javax.swing.JPanel panelThemAnh;
+    private javax.swing.JRadioButton radioNam;
+    private javax.swing.JRadioButton radioNu;
+    private javax.swing.JTable tableHienThi;
+    private javax.swing.JTextField txtCmnd;
+    private javax.swing.JTextField txtDiaChi;
+    private javax.swing.JTextField txtEmail;
+    private javax.swing.JTextField txtMaNhanVien;
+    private javax.swing.JTextField txtSdt;
+    private javax.swing.JTextField txtTenNhanVien;
+    private javax.swing.JTextField txtTenTaiKhoan;
+    // End of variables declaration//GEN-END:variables
+}
